Fixed timestamp of pointer grab.
In the previous version, ScreenView._grabPointerTime was later than
timestamp of ButtonPress event. motif did not work because of it. As
seen in lib/Xm/DragC.c in the motif source tree, motif gives a last
event timestamp to XGrabPointer().

When a user pressed the left button, the X server and motif worked in
this order:

  1. The X server sent a ButtonPress event.
  2. The X server updated the _grabPointerTime in the ScreenView.
  3. motif called XGrabPointer() with the timestamp given in the step 1.
  4. The X server called ScreenView.processGrabPointerRequest().

In the step 4, the X server responsed "invalid time" in comparing the
timestamp at the step 2 with the timestamp at the step 3 (which is same
as the timestamp at the step 1).

This patch fixes the X server to use same timestamp in the step 1 and 2.