Use VIRTUAL_KEYBOARD as fallback keyboard type
Some devices don't have keyboard with full type, and the
Android OS uses the KeyCharacterMap.BUILT_IN_KEYBOARD
(value is 0) as the fallback keyboard type. If the
phone/tablet doesn't have built in keyboard, the returned
KeyCharacterMap is empty, and it will causes the
ArrayOutOfBounds exception when initializing keyboard
mapping. So this CL uses VIRTUAL_KEYBOARD as the fallback
keyboard type if the phone/tablet doesn't hava a FULL
type keyboard.

Signed-off-by: utzcoz <utzcoz@outlook.com>